requirements.txt
CacheControl==0.12.5
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.4
django-rest-framework==0.1.0
djangorestframework==3.9.0
drf-yasg==1.12.1
firebase-admin==2.16.0
google-api-core==1.8.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.18.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
msgpack==0.6.1
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.9
requests==2.21.0
rsa==4.0
ruamel.yaml==0.15.83
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
